Alignement on netbsd types, part 1
The following types are modified (old -> new):
 * _BSD_USECONDS_T_ int       -> unsigned int
 * __socklen_t      __int32_t -> __uint32_t
 * blksize_t        uint32_t  -> int32_t
 * rlim_t           uint32_t  -> uint64_t
On ARM:
 * _BSD_CLOCK_T_    int       -> unsigned int
On Intel:
 * _BSD_CLOCK_T_    int       -> unsigned long

bin/cat is also updated in order to fix warnings.

_BSD_TIMER_T_ has still to be aligned.

+++ include/dirent.h
@@ -113,7 +113,7 @@ DIR *__opendir2(const char *, int) __RENAME(__opendir230);
 int scandir(const char *, struct dirent ***,
     int (*)(const struct dirent *), int (*)(const void *, const void *))
     __RENAME(__scandir30);
-ssize_t getdents(int, char *, size_t) __RENAME(__getdents30);
+int getdents(int, char *, size_t) __RENAME(__getdents30);
 int alphasort(const void *, const void *) __RENAME(__alphasort30);
 #endif
 #endif /* defined(_NETBSD_SOURCE) */
